requirements.txt
qrcode
pillow
python-dotenv